+++ src/__tests__/rule-tester/RuleTester.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom/extend-expect';
 
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 
-import { mockGetWorkItem } from '../../__mocks__/azure-devops-extension-api/Wit';
+import { mockGetWorkItem } from '../../__mocks__/azure-devops-extension-api/WorkItemTracking';
 import { mockGetConfiguration } from '../../__mocks__/azure-devops-extension-sdk';
 import {
   getWorkItem,
@@ -21,7 +21,7 @@ describe('RuleTester', () => {
   beforeEach(() => {
     jest.clearAllMocks();
     jest.spyOn(WorkItemService.prototype, 'getWorkItemTypes').mockResolvedValue(getWorkItemTypes());
-    jest.spyOn(StorageService.prototype, 'getData').mockResolvedValue([]);
+    jest.spyOn(StorageService.prototype, 'getRuleDocuments').mockResolvedValue([]);
   });
 
   it('should render default', async () => {
@@ -185,8 +185,6 @@ describe('RuleTester', () => {
     const testButton = screen.getByRole('button', { name: 'Test' });
     fireEvent.click(testButton);
 
-    await waitFor(() =>
-      screen.getAllByText(/n error occurred while testing the work item:/)
-    );
+    await waitFor(() => screen.getAllByText(/n error occurred while testing the work item:/));
   });
 });