+++ src/settings-panel/module.tsx
@@ -1,0 +1,6 @@
+import './index.scss';
+
+import showRootComponent from '../shared-ui/showRootComponent';
+import SettingsPanel from './SettingsPanel';
+
+showRootComponent(<SettingsPanel />, 'settings-panel-container');