+++ src/shared-ui/showRootComponent.tsx
@@ -1,3 +1,5 @@
+/* istanbul ignore file */
+
 import { render } from 'react-dom';
 
 function showRootComponent(component: React.ReactElement<any>, target: string): void {