+++ CHANGELOG.md
@@ -2,16 +2,24 @@
 
 ## 1.2.0 (2022-01-XX)
 
-### 🛠️ Maintenance (1)
+### 🛠️ Maintenance (2)
 
 - Update azext and implement new changes
+
   - Changed in [PR#21 - Update azext](https://github.com/joachimdalen/azdevops-auto-state/pull/21)
 
-### 🚀 Features (1)
+- Update dependencies
+
+### 🚀 Features (2)
 
 - Added the ability to disable rules
+
   - Suggested in [GH#22 - Add ability to disable rule](https://github.com/joachimdalen/azdevops-auto-state/issues/22)
   - Added in [PR#26 - Add ability to disable rules](https://github.com/joachimdalen/azdevops-auto-state/pull/26)
+
+- Add setting to be able to scope work item types to current process
+  - Suggested in [GH#23 - Add setting to be able to limit work item types to current process](https://github.com/joachimdalen/azdevops-auto-state/issues/23)
+  - Added in [PR#27 - Add settings module](https://github.com/joachimdalen/azdevops-auto-state/pull/27)
 
 ## 📦 Module changes
 
@@ -20,6 +28,13 @@
 #### `rule-modal@1.2.0`
 
 - Fixed UI showing/flashing before loading was completed
+
+### 📣 Enhancements (1)
+
+#### `rule-modal@1.2.0`
+
+- Added labels to toggle states to indicate the current state
+  - Improved in [PR#27 - Add settings module](https://github.com/joachimdalen/azdevops-auto-state/pull/27)
 
 ---